requirements.txt
streamlit>=1.36
python-dotenv>=1.0
google-generativeai>=0.8
openai>=1.40
pypdf>=4.2
python-docx>=1.1
